Update tests with new Jest version
