+++ lib/assets/javascripts/new-dashboard/store/modules/maps/carto/featured-favorited-maps.js
@@ -1,8 +1,5 @@
-import CartoNode from 'carto-node';
 import toObject from 'new-dashboard/utils/to-object';
 import { defaultParams as filtersDefaultParams } from 'new-dashboard/core/configuration/filters';
-
-const client = new CartoNode.AuthenticatedClient();
 
 const featuredFavoritedMaps = {
   namespaced: true,
@@ -54,7 +51,7 @@ const featuredFavoritedMaps = {
 
       context.commit('setFetchingState');
 
-      client.getVisualization('',
+      context.rootState.client.getVisualization('',
         favoritedParams,
         function (err, _, data) {
           if (err) {
