+++ lib/assets/javascripts/new-dashboard/store/modules/user/index.js
@@ -1,9 +1,6 @@
-import CartoNode from 'carto-node';
 import getCARTOData from 'new-dashboard/store/utils/getCARTOData';
 
 import * as getters from './getters';
-
-const client = new CartoNode.AuthenticatedClient();
 
 // Data coming from Static Page Loading,
 // which requests user and config data from
@@ -51,7 +48,7 @@ const user = {
   },
   actions: {
     updateData (context) {
-      client.getConfig(function (err, _, data) {
+      context.rootState.client.getConfig(function (err, _, data) {
         if (err) {
           return;
         }
