+++ lib/assets/test/spec/new-dashboard/unit/specs/components/NavigationBar.spec.js
@@ -53,6 +53,6 @@ describe('NavigationBar.vue', () => {
       localVue
     });
 
-    expect(NavigationBar).toMatchSnapshot();
+    expect(navigationBar).toMatchSnapshot();
   });
 });
