+++ lib/assets/test/spec/new-dashboard/unit/specs/pages/Home/RecentSection.spec.js
@@ -4,8 +4,8 @@ import visualizations from '../../fixtures/visualizations';
 import datasets from '../../fixtures/datasets';
 
 const recentContentList = [
-  ...visualizations.visualizations[0],
-  ...datasets.visualizations[0]
+  visualizations.visualizations[0],
+  datasets.visualizations[0]
 ];
 
 describe('RecentSection.vue', () => {
