+++ lib/assets/test/spec/new-dashboard/unit/specs/store/maps/feature-favorited-maps.spec.js
@@ -1,4 +1,5 @@
 import visualization from '../../fixtures/visualizations';
+import datasets from '../../fixtures/datasets';
 import FeaturedFavoritedMapsStore from 'new-dashboard/store/modules/maps/carto/featured-favorited-maps';
 import toObject from 'new-dashboard/utils/to-object';
 import { testAction } from '../helpers';
@@ -7,6 +8,19 @@ jest.mock('carto-node');
 
 const mutations = FeaturedFavoritedMapsStore.mutations;
 const actions = FeaturedFavoritedMapsStore.actions;
+
+const getVisualizationMock = function getVisualization (_, params, callback) {
+  let data = visualization;
+  if (params.types === 'table') {
+    data = datasets;
+  }
+  if (typeof params.order === 'boolean') {
+    const err = { error: "Wrong 'order' parameter value. Valid values are one of [:updated_at, :size, :mapviews, :likes]" };
+    callback(err, null, null);
+  } else {
+    callback(null, null, data);
+  }
+};
 
 describe('FeaturedFavoritedMapsStore', () => {
   describe('mutations', () => {
@@ -69,7 +83,7 @@ describe('FeaturedFavoritedMapsStore', () => {
 
   describe('actions', () => {
     describe('fetchMaps', () => {
-      let state;
+      let state, rootState;
 
       beforeEach(() => {
         state = {
@@ -85,6 +99,12 @@ describe('FeaturedFavoritedMapsStore', () => {
             only_liked: true
           }
         };
+
+        rootState = {
+          client: {
+            getVisualization: getVisualizationMock
+          }
+        };
       });
 
       it('success', done => {
@@ -92,7 +112,7 @@ describe('FeaturedFavoritedMapsStore', () => {
           { type: 'setFetchingState' },
           { type: 'setFeaturedFavoritedMaps', payload: visualization }
         ];
-        testAction({action: actions.fetchMaps, state, expectedMutations, done});
+        testAction({ action: actions.fetchMaps, state, rootState, expectedMutations, done });
       });
 
       it('errored', done => {
@@ -102,7 +122,7 @@ describe('FeaturedFavoritedMapsStore', () => {
           { type: 'setFetchingState' },
           { type: 'setRequestError', payload: err }
         ];
-        testAction({action: actions.fetchMaps, state, expectedMutations, done});
+        testAction({ action: actions.fetchMaps, state, rootState, expectedMutations, done });
       });
     });
   });
