+++ lib/assets/test/spec/new-dashboard/unit/specs/store/user/user.spec.js
@@ -57,6 +57,19 @@ describe('UserStore', () => {
       });
 
       it('success', done => {
+        const rootState = {
+          client: {
+            getConfig (callback) {
+              callback(null, null, {
+                user_data: {
+                  email: 'example@carto.com',
+                  username: 'carto'
+                }
+              });
+            }
+          }
+        };
+
         const newConfigData = {
           email: 'example@carto.com',
           username: 'carto'
@@ -64,7 +77,8 @@ describe('UserStore', () => {
         const expectedMutations = [
           { type: 'setUserData', payload: newConfigData }
         ];
-        testAction({ action: actions.updateData, state, expectedMutations, done });
+
+        testAction({ action: actions.updateData, state, rootState, expectedMutations, done });
       });
     });
